+++ packages/cli/src/clonebins_cli/__init__.py
@@ -4,4 +4,4 @@ from clonebins_cli.main import app
 
 __all__ = ["app"]
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"